+++ cacheutilslibrary/src/main/java/com/lifeofcoding/cacheutlislibrary/GsonHelper.java
@@ -4,7 +4,6 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonToken;
-
 import java.io.IOException;
 import java.util.Date;
